+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Version 3.0.0 - 2021-10-14
 
 - Removed all support for batches. Use datasets instead.
-- Added UpdateDocumentOptions
+- Added UpdateDocumentOptions to updateDocument
 - Added datasetId, groundTruth and retentionInDays to UpdateDocumentOptions
 
 ## Version 2.0.0 - 2021-08-31